discord-bot/requirements.txt
discord.py>=2.3,<3
anthropic>=0.39,<1
python-dotenv>=1.0,<2
